refactor locklocation into its own file
